started refactoring the test node
